fix small bugs

+++ website/resources/views/poles/competitions/show.blade.php
@@ -135,12 +135,14 @@
 						<a class="btn btn-primary btn-rounded" href="/poles/competitions/{{ $compet->id }}/edit">Modifier cette
 							compétition</a>
 					</div>
+			@endcan
+			@can('delete', $compet)
 					<div class="text-center" style="margin-top:25px; margin-bottom:25px;">
 						<a class="btn btn-primary btn-rounded"
 							href="/poles/competitions/{{ $compet->id }}/destroy">Supprimer</a>
 					</div>
-				</div>
 			@endcan
+				</div>
 
 			<hr>
